+++ packages/runtime/adapters/monitoring-adapters/monitoring-aws-cloudwatch/src/index.ts
@@ -26,6 +26,7 @@ const createMonitoringImplementation = (
     getMetrics: () => monitoringContext.monitoringBase.getMetrics(),
     clearMetrics: () => monitoringContext.monitoringBase.clearMetrics(),
     publish: monitoringPublish.bind(null, log, monitoringContext),
+    custom: monitoringContext.monitoringBase.custom,
   }
 }