+++ packages/runtime/adapters/readmodel-adapters/readmodel-postgresql/src/connect.ts
@@ -7,6 +7,7 @@ import type {
   OmitObject,
 } from './types'
 
+const MAX_DISCONNECT_TIME = 3000
 const connect: CurrentConnectMethod = async (imports, pool, options) => {
   let {
     tablePrefix,
@@ -70,7 +71,6 @@ const connect: CurrentConnectMethod = async (imports, pool, options) => {
             connectionErrorsMap.get(connection)?.push(error)
           })
           await connection.connect()
-          await connection.query('SELECT 0 AS "defunct"')
         } catch (error) {
           connectionErrorsMap.get(connection)?.push(error)
         }
@@ -99,16 +99,26 @@ const connect: CurrentConnectMethod = async (imports, pool, options) => {
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         pool.connection = null!
       }
+
+      affectedReadModelOperationsSet.clear()
+      let timeout: NodeJS.Timeout | null = null
       try {
-        affectedReadModelOperationsSet.clear()
-        await connection.end()
-      } catch (err) {}
+        await Promise.race([
+          new Promise((resolve) => {
+            timeout = setTimeout(resolve, MAX_DISCONNECT_TIME)
+          }),
+          connection.end(),
+        ])
+      } catch (err) {
+      } finally {
+        if (timeout != null) {
+          clearTimeout(timeout)
+        }
+      }
+
       throw summaryError
     }
   }
-
-  const initialConnection = await establishConnection()
-  await maybeThrowConnectionErrors(initialConnection)
 
   const inlineLedgerRunQuery: InlineLedgerRunQueryMethod = async (
     sql,
@@ -184,7 +194,8 @@ const connect: CurrentConnectMethod = async (imports, pool, options) => {
     schemaName: databaseName,
     tablePrefix,
     inlineLedgerRunQuery,
-    connection: initialConnection,
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    connection: null!,
     activePassthrough: false,
     buildMode,
     useSqs,